requirements.txt
pandas
numpy
mlflow
scikit-learn
faker
pyyaml 
python-box
-e .